Split main.js to several files
